+++ weblate/trans/models/translation.py
@@ -1468,9 +1468,23 @@ class Translation(
             existing = set(self.unit_set.values_list("context", flat=True))
         else:
             existing = set(self.unit_set.values_list("context", "source"))
+            # Iterate over list to copy set that will be changed
+            for ex_context, ex_source in list(existing):
+                if is_plural(ex_source):
+                    # Include unpluralized string as well as it does not work in most formats
+                    existing.add((ex_context, split_plural(ex_source)[0]))
+
         for _set_fuzzy, unit in store.iterate_merge(fuzzy, only_translated=False):
             idkey = unit.context if has_template else (unit.context, unit.source)
             if idkey in existing:
+                skipped += 1
+                continue
+            # Check for singular in existing units as well, this does not work well in most formats
+            if (
+                not has_template
+                and is_plural(unit.source)
+                and (unit.context, split_plural(unit.source)[0]) in existing
+            ):
                 skipped += 1
                 continue
             self.add_unit(
@@ -2092,7 +2106,7 @@ class Translation(
         skip_existing: bool = False,
     ) -> None:
         component = self.component
-        extra = {}
+        extra = []
         if isinstance(source, str):
             source = [source]
         if len(source) > 1 and not component.file_format_cls.supports_plural:
@@ -2120,8 +2134,20 @@ class Translation(
         if context:
             component.file_format_cls.validate_context(context)
         if not component.has_template():
-            extra["source"] = join_plural(source)
-        if not auto_context and self.unit_set.filter(context=context, **extra).exists():
+            source_query = Q(source=join_plural(source))
+            # Validate non-pluralized strings against pluralized ones because having same with the same sources
+            # does not work for most of the formats
+
+            # Look for plural string with the same singular (also
+            source_query |= Q(source__startswith=join_plural([source[0], ""]))
+
+            # Look for singular string with the same text
+            if len(source) > 1:
+                source_query |= Q(source=source[0])
+
+            extra.append(source_query)
+
+        if not auto_context and self.unit_set.filter(*extra, context=context).exists():
             raise ValidationError(gettext("This string seems to already exist."))
         # Avoid using source translations without a filename
         if not self.filename: